Release 0-1-0 package

+++ CardMagnifier/MonoBehaviors/CardEnlarger.cs
@@ -561,8 +561,8 @@ namespace CardMagnifier.MonoBehaviors
                 mapEmbiggenerScale = cameraOrthoSizeGameplay / cameraOrthoSizeCardPick;
                 screenResolutionScale = Mathf.Min
                 (
-                    currentResolution.x / defaultResolution.x,
-                    currentResolution.y / defaultResolution.y
+                    (float)currentResolution.x / (float)defaultResolution.x,
+                    (float)currentResolution.y / (float)defaultResolution.y
                 );
             }
